Create replace_dots_with_underscore
Remove all the dots in the filename except the one before the extension and replace them with underscore. Preserves the dot before the extension and also preserves the extension. Works with any extension and any mix of extensions in the filenames.
eg. change "Image 2022-07-08 at 11.09.24 PM.png" to "Image 2022-07-08 at 11_09_24 PM.png"